requirements.txt
IPy==0.82a
Pillow==2.8.1
miniupnpc==1.9
psutil==2.2.1
bitcoin==1.1.27
pycountry==1.10
pyelliptic==1.5.5
pystun==0.1.0
python-gnupg==0.3.7
python-obelisk==0.1.3
qrcode==5.1
requests==2.5.1
tornado==4.0.2
Twisted==14.0.2
rfc3986==0.2.1
dnschain==0.1.0
pyee==0.1.0
